requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.12
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
pafy==0.5.5
psycopg2==2.8.6
pytube3==9.6.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.7.4.3
whitenoise==5.2.0
youtube-dl==2021.4.26
